Refactor: Improve data logic and component structure
This commit introduces several refactoring improvements to enhance code organization and maintainability.

- Refactor database logic in `src/db/index.js` by creating a shared `executeSql` promise wrapper. This reduces code duplication and simplifies CRUD functions.

- Extract `SkiaCard` component into its own file `src/components/Card.js` from the main editor screen.

- Move data-related state and logic from `FlowEditorScreen.js` to a dedicated custom hook `useFlowData.js`. This includes functions for node/edge manipulation and hierarchy navigation.
